i965/drv: Always guess the expected FOURCC
We can simplify the logic in i965_CreateSurfaces2 to behave exactly like iHD, meaning that we can just have an ugly switch and then let the user override the surface later on. An invalid surface FOURCC will error out safely anyways.

I've left out some formats that might be worth hooking up to the driver in the future, still need to have a think about this.